Use a hash as storage for private queues
Using lists as the private backup for queues has a major flaw: to remove
a job from the backup queue the serialized value must match exactly.
There are at several situations where the original job and the
serialized backup may not match:

1. The job was enqueued by Sidekiq, which doesn't serialize JSON with
ordered keys.
2. A value was added to the job during processing, i.e. a `pid`

This commit changes the backup from a list to a hash. The hash uses the
`jid` as a key, which guarantees that it can be found later when it is
time to clean up the job.